refactor(cli): extract doctor+add into commands/setup.ts (5.0-alpha setup Phase A)
Fourth cut of the 5.0-alpha cli.ts god-module split. Creates
src/commands/setup.ts and moves the two self-contained setup-cluster handlers
cmdDoctor and cmdAdd into it.

- Phase A only: cmdSetup + cmdInit intentionally stay in cli.ts (cmdSetup
  orchestrates cmdInstall/cmdLogin/cmdCheck, still inline — moving it now would
  force a commands -> cli.ts back-import). They join setup.ts in Phase B once
  those clusters are extracted.
- cmdDoctor + cmdAdd exported; cli.ts imports them back. COMMANDS dispatch
  table unchanged.
- Prunes now-unused cli.ts imports (runDoctor from doctor.js; provisionService/
  listAvailableServices/getServiceInfo from provision.js) — doctor+add were
  their only users.

cli.ts: 4422 -> 4322 lines (-100; -2008 cumulative across PRs 1-4, from 6330).
Typecheck, eslint (0 errors), full test suite (2580), and prettier all green.
