+++ src/cli.ts
@@ -68,7 +68,6 @@ import {
   updateSkillsLock,
   updateCliLock,
 } from "./lock.js";
-import { provisionService, listAvailableServices, getServiceInfo } from "./provision.js";
 import { cmdFix } from "./fix.js";
 import { promptConfirm } from "./utils/prompt.js";
 import { c } from "./utils/colors.js";
@@ -96,12 +95,12 @@ import { cmdSbom } from "./commands/sbom.js";
 import { cmdSecurity } from "./commands/security.js";
 import { cmdSecrets } from "./commands/secrets.js";
 import { cmdUpgrade, selfUpgrade } from "./commands/upgrade.js";
+import { cmdDoctor, cmdAdd } from "./commands/setup.js";
 import { cmdAuth } from "./commands/auth.js";
 import { cmdAudit } from "./commands/audit.js";
 import { cmdMcp } from "./commands/mcp.js";
 import { cmdHooks } from "./commands/hooks.js";
 import { resolveAllAuth } from "./service-auth.js";
-import { runDoctor } from "./doctor.js";
 import { detectStack } from "./stack-detector.js";
 import { generateToml, parseEnvTemplateKeys } from "./toml-generator.js";
 import { vaultMeta, detectSecretStore } from "./vault-meta.js";
@@ -2051,105 +2050,6 @@ async function offerFirstInstallPrescan(): Promise<void> {
     console.log(`\n${c.dim}Full report:${c.reset} ${report.summaryPath}`);
   }
   console.log();
-}
-
-async function cmdDoctor(): Promise<boolean> {
-  console.log(`${c.bold}${c.cyan}kit doctor${c.reset}`);
-  console.log(`${c.dim}${"─".repeat(50)}${c.reset}\n`);
-
-  let config: ReturnType<typeof Object.create> = {};
-  try {
-    config = await loadConfig(resolveConfigPath());
-  } catch {
-    // Doctor works even without .kit.toml — skip config-dependent checks
-  }
-
-  const result = await runDoctor(config, process.cwd());
-
-  for (const check of result.checks) {
-    const icon =
-      check.status === "pass"
-        ? `${c.green}✓${c.reset}`
-        : check.status === "warn"
-          ? `${c.yellow}⚠${c.reset}`
-          : check.status === "fail"
-            ? `${c.red}✗${c.reset}`
-            : `${c.dim}–${c.reset}`;
-    console.log(`  ${icon} ${check.name}  ${c.dim}${check.detail}${c.reset}`);
-  }
-
-  console.log();
-  console.log(`${c.dim}${"─".repeat(50)}${c.reset}`);
-
-  const summaryParts: string[] = [];
-  if (result.passed > 0) summaryParts.push(`${c.green}${result.passed} passed${c.reset}`);
-  if (result.warnings > 0)
-    summaryParts.push(
-      `${c.yellow}${result.warnings} warning${result.warnings === 1 ? "" : "s"}${c.reset}`,
-    );
-  if (result.failed > 0) summaryParts.push(`${c.red}${result.failed} failed${c.reset}`);
-  if (summaryParts.length === 0) summaryParts.push(`${c.dim}no checks ran${c.reset}`);
-
-  console.log(`\n  ${summaryParts.join(" · ")}\n`);
-
-  return result.failed === 0;
-}
-
-async function cmdAdd(): Promise<boolean> {
-  const serviceName = process.argv[3];
-
-  if (!serviceName) {
-    console.log(`${c.bold}${c.cyan}Available services:${c.reset}\n`);
-
-    const services = listAvailableServices();
-    for (const svc of services) {
-      const info = getServiceInfo(svc);
-      if (info) {
-        console.log(`  ${c.green}${svc}${c.reset}  ${c.dim}${info.description}${c.reset}`);
-        console.log(`    ${c.dim}Requires: ${info.tools.join(", ")}${c.reset}`);
-      }
-    }
-
-    console.log();
-    console.log(`${c.dim}Usage: kit add <service>${c.reset}`);
-    console.log(`${c.dim}Example: kit add stripe/payments${c.reset}`);
-    console.log();
-    return false;
-  }
-
-  console.log(`${c.bold}${c.cyan}Provisioning ${serviceName}...${c.reset}\n`);
-
-  const projectPath = process.cwd();
-  const projectName = process.cwd().split("/").pop();
-
-  const result = await provisionService(serviceName, projectPath, projectName);
-
-  if (result.success) {
-    console.log(`  ${c.green}✓${c.reset} ${result.message}`);
-
-    if (result.secrets && Object.keys(result.secrets).length > 0) {
-      console.log();
-      console.log(`  ${c.dim}Added secrets to .env.local:${c.reset}`);
-      for (const key of Object.keys(result.secrets)) {
-        console.log(`    ${c.cyan}${key}${c.reset}`);
-      }
-    }
-
-    if (result.config) {
-      console.log();
-      console.log(`  ${c.dim}Updated skills-lock.json${c.reset}`);
-    }
-
-    console.log();
-    return true;
-  } else {
-    console.log(`  ${c.red}✗${c.reset} ${result.message}`);
-    if (result.error) {
-      console.log(`  ${c.dim}Error: ${result.error}${c.reset}`);
-    }
-    console.log();
-    return false;
-  }
 }
 
 type CiFormat = "github" | "gitlab" | "json" | "text";
